+++ Sources/BraintreeVenmo/BTVenmoAppSwitchReturnURL.h
@@ -3,6 +3,7 @@
 extern NSString *const BTVenmoAppSwitchReturnURLErrorDomain;
 
 typedef NS_ENUM(NSUInteger, BTVenmoAppSwitchReturnURLState) {
+    BTVenmoAppSwitchReturnURLStateSucceededWithPaymentContext,
     BTVenmoAppSwitchReturnURLStateSucceeded,
     BTVenmoAppSwitchReturnURLStateFailed,
     BTVenmoAppSwitchReturnURLStateCanceled,
@@ -48,6 +49,11 @@ typedef NS_ENUM(NSUInteger, BTVenmoAppSwitchReturnURLState) {
  The username from the return URL.
 */
 @property (nonatomic, copy, readonly) NSString *username;
+
+/**
+ The payment context ID from the return URL.
+ */
+@property (nonatomic, copy, readonly) NSString *paymentContextID;
 
 /**
  If the return URL's state is BTVenmoAppSwitchReturnURLStateFailed, the error returned from Venmo via the app switch.